Add storybook example for ref panel
